Single element vector fix for opaque pointers
If writer adaptor is using typed pointers it can generate an
VCSingleElementVector attribute with a number of nested pointers.
In reader values with such attribute should not be restored into SEVs in
case of opaque pointers. In this mode only top level pointer is present
and all the nested ones are lost.