Always restore SEV for opaque pointers
When an opaque pointer argument or global has a SEV attribute, assume
that it has to be restored as <1 x ptr> regardless of the value of the
attribute (inner pointers count). This value should always be zero
when opaque pointers are used on the writer side, but can be non-zero
otherwise. In such cases the original type info is lost and the only
thing left is to restore SEV and hope it won't crash.