Update LLVM version check for IIT_PTR_TO_* intrinsics
`IIT_PTR_TO_ELT` and `IIT_PTR_TO_ARG` intrinsics were removed from LLVM 17.
https://github.com/llvm/llvm-project/commit/4136e08f14ae9043b46bbe0a101e8609164d1be5 https://github.com/llvm/llvm-project/commit/253a52988fe37887ba1f7741271d602083fe7dd7

There's a check for `VC_INTR_LLVM_VERSION_MAJOR < 17` later on in a switch (code link), but in the enum definiton it's set to "< 18". This gives a warning when building with `LLVM_PREFERRED_VERSION` set to 17, as the values are still defined but not used in the switch. Using `-Werror=switch` makes the build fail, as expected.

This change sets the condition to "< 17" to be in line with the one defined in the switch.